Create grid.service
Temporary changes- not working
