+++ src/app/customize/grid-menu/grid-menu.component.html
@@ -1,5 +1,5 @@
 <div style="background: transparent">
 
-    <div [p3x-compile]="getWidgets().getWidget(item.id).menuTemplate" [p3x-compile-ctx]="this" [item]="item"></div>
+    <div [p3x-compile]="getWidgets().getWidget(item.id).menuTemplate" [uid]="item.uid"></div>
 
 </div>
