+++ src/app/grid/grid-content.component.html
@@ -1,6 +1,6 @@
 
 <div style="background: transparent; height: 100%; width: 100%;" class="child">
 
-  <div [p3x-compile]="getWidgets().getWidget(item.id).template" [p3x-compile-ctx]="this" [item]="item"></div>
+  <div [p3x-compile]="getWidgets().getWidget(item.id).template" [uid]="item.uid"></div>
 
 </div>
